% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.r
\name{get_token}
\alias{get_token}
\alias{new_token}
\alias{get_domain}
\alias{load_token_and_domain}
\title{Token management}
\usage{
get_token()

new_token(token, domain)

get_domain()

load_token_and_domain()
}
\value{
A canvas token
}
\description{
Token management
}
